refactor(#1459): extract PermissionProtocol + types module (Phase 3)
- Create services/permissions/types.py with 8 extracted types:
  ConsistencyLevel, ConsistencyMode, ConsistencyRequirement, WriteResult,
  CheckResult, TraversalStats, GraphLimits, GraphLimitExceeded
- Create services/protocols/permission.py with @runtime_checkable
  PermissionProtocol defining 6 Zanzibar APIs (check, check_bulk, write,
  delete, expand, list_objects)
- Update rebac_manager_enhanced.py to import from types.py with
  backward-compat re-exports
- Add PermissionProtocol to protocols/__init__.py exports
